+++ components/experimental/external-component-loader/src/externalComponentLoader/index.ts
